Extracting out dialog logic from script.js
